Structure as a python module
